refactor: share e2e analytics-mock fixture across sites (#1637)
